Краткое отображение количества ресурсов

+++ Assets/Scripts/Stash.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Stash
 {
-    public static int CoinsCount;
-    public static int[] ResourceCount = new int[7];
+    public static long CoinsCount;
+    public static long[] ResourceCount = new long[7];
     public static int[] ResourcePrices = new int[7] { 1, 3, 5, 10, 15, 30, 60};
 }
